remove part1_native and use flag for split kernel
